+++ test/test_states.jl
@@ -1,12 +1,12 @@
 
 
 using DecoratedParticles, StaticArrays, Test, LinearAlgebra
-using DecoratedParticles: State, DState
+using DecoratedParticles: PState, VState
 DP = DecoratedParticles
 
 ##
 
-X = State( rr = randn(SVector{3, Float64}) )
+X = PState( rr = randn(SVector{3, Float64}) )
 
 SYMS = DP._syms(X)
 println(@test SYMS == (:rr,))
@@ -14,22 +14,22 @@ println(@test SYMS == (:rr,))
 TT = DP._tt(X)
 println(@test( TT == Tuple{SVector{3, Float64}} ))
 
-TDX = DP.dstate_type(X)
-println(@test TDX == DState{NamedTuple{(:rr,), Tuple{SVector{3, Float64}}}})
+TDX = DP.vstate_type(X)
+println(@test TDX == VState{NamedTuple{(:rr,), Tuple{SVector{3, Float64}}}})
 
 
-dX = DState(X)
-println(@test typeof(dX) == DP.dstate_type(X))
+dX = VState(X)
+println(@test typeof(dX) == DP.vstate_type(X))
 println(@test dX.rr == X.rr)
 
 
-cTDX = DP.dstate_type(0.0im, X)
-println(@test cTDX == DState{NamedTuple{(:rr,), Tuple{SVector{3, ComplexF64}}}})
+cTDX = DP.vstate_type(0.0im, X)
+println(@test cTDX == VState{NamedTuple{(:rr,), Tuple{SVector{3, ComplexF64}}}})
 
 cdX = complex(dX)
-println(@test( cdX == DState(rr = X.rr .+ 0im) ))
+println(@test( cdX == VState(rr = X.rr .+ 0im) ))
 println(@test( real(cdX) == dX ))
-println(@test( imag(cdX) == DState(rr = zero(SVector{3, Float64})) ))
+println(@test( imag(cdX) == VState(rr = zero(SVector{3, Float64})) ))
 
 # not sure how to test this, but at least it should work:
 @show rand(TDX)
@@ -41,15 +41,15 @@ println(@test( imag(cdX) == DState(rr = zero(SVector{3, Float64})) ))
 @show zero(X)
 
 @info("arithmetic")
-println(@test( X + dX == State(rr = X.rr + dX.rr) ))
-println(@test( X - dX == State(rr = X.rr - dX.rr) ))
-println(@test( dX + cdX == DState(rr = dX.rr + cdX.rr) ))
-println(@test( dX - cdX == DState(rr = dX.rr - cdX.rr) ))
+println(@test( X + dX == PState(rr = X.rr + dX.rr) ))
+println(@test( X - dX == PState(rr = X.rr - dX.rr) ))
+println(@test( dX + cdX == VState(rr = dX.rr + cdX.rr) ))
+println(@test( dX - cdX == VState(rr = dX.rr - cdX.rr) ))
 
 a = randn() 
-println(@test( a * dX == DState(rr = a * dX.rr) ))
-println(@test( dX * a == DState(rr = a * dX.rr) ))
-println(@test( -dX == DState(rr = -dX.rr) ))
+println(@test( a * dX == VState(rr = a * dX.rr) ))
+println(@test( dX * a == VState(rr = a * dX.rr) ))
+println(@test( -dX == VState(rr = -dX.rr) ))
 
 println(@test( dot(dX, cdX) == dot(dX.rr, cdX.rr) ))
 println(@test( dot(cdX, dX) == dot(cdX.rr, dX.rr) ))
@@ -77,8 +77,10 @@ function bm_copy!(Y, X, a)
    return Y
 end
 
-Xs = [ rand(PositionState{Float64}) for _=1:100 ]
-Ys = [ zero(PositionState{Float64}) for _=1:100 ]
+PositionStateF64 = typeof(PState(rr = zero(SVector{3, Float64})))
+
+Xs = [ rand(PositionStateF64) for _=1:100 ]
+Ys = [ zero(PositionStateF64) for _=1:100 ]
 a = rand() 
 let Ys = Ys, Xs = Xs, a = a
    bm_copy!(Ys, Xs, a)
@@ -86,3 +88,4 @@ let Ys = Ys, Xs = Xs, a = a
    @show nalloc 
    println(@test nalloc == 0) 
 end
+
